Results from specify location, grid and choose plant tied together
Updated how the viewmodels are accessed and created a viewmodel for the
current garden to keep track of this before it is saved in the garden
viewmodel. To do so we needed the fragment:fragment_ktx library.

Select Plant text is not yet shown and might be deleted later but for
now it is in the string resources.

Gardens are represented by:

* Location
* Name
* Plants
* Grid tiles (tileIds)

Moved everything regarding the recyclerview into ChoosePlantFragment to
provide the Viewholder with data for adding items to the current garden viewmodel